+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+#### 1.0.1
+* Fixed missing flush instruction in to_array() method
+* Fixed representation for char columns in to_array() method
+* Fixed internal function name typo in _replace_by_dataframe()
+* Improved error handling for invalid row item types when reading CSV files
+
 #### 1.0.0 
 * Open source release